bot/requirements.txt
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.10.4
asyncpg==0.30.0
certifi==2025.7.14
click==8.2.1
fastapi==0.116.1
GeoAlchemy2==0.17.1
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.25.2
idna==3.10
jinja2==3.1.0
Mako==1.3.10
MarkupSafe==3.0.2
numpy==2.3.1
packaging==25.0
Pillow>=10.0.0
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.9
python-telegram-bot==20.7
pytz==2025.2
PyYAML==6.0.2
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzlocal==5.3.1
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
